# docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_Slazzzer_golang-todoapp_internal_core_transport_http_types.Nullable-bool:
    properties:
      set:
        type: boolean
      value:
        type: boolean
    type: object
  github_com_Slazzzer_golang-todoapp_internal_core_transport_http_types.Nullable-string:
    properties:
      set:
        type: boolean
      value:
        type: string
    type: object
  internal_core_transport_http_probes.ProbeStatusResponse:
    properties:
      status:
        description: 'Статус: ok, ready или not ready'
        example: ok
        type: string
    type: object
  internal_features_statistics_transport_http.GetStatisticsResponse:
    properties:
      task_completed_rate:
        description: Процент завершённых (0–100) или null
        example: 70
        type: number
      tasks_average_completion_time:
        description: Среднее время выполнения или null
        example: 24h30m15s
        type: string
      tasks_completed:
        description: Завершено задач за период
        example: 7
        type: integer
      tasks_created:
        description: Создано задач за период
        example: 10
        type: integer
    type: object
  internal_features_tasks_transport_http.CreateTaskRequest:
    properties:
      author_user_id:
        description: ID автора задачи
        example: 1
        type: integer
      description:
        description: Описание (опционально)
        example: 2 литра
        maxLength: 1000
        minLength: 1
        type: string
      title:
        description: Заголовок (1–100 символов)
        example: Купить молоко
        maxLength: 100
        minLength: 1
        type: string
    required:
    - author_user_id
    - title
    type: object
  internal_features_tasks_transport_http.CreateTaskResponse:
    properties:
      author_user_id:
        description: ID автора (из JWT)
        example: 1
        type: integer
      completed:
        description: Выполнена ли задача
        example: false
        type: boolean
      completed_at:
        description: Дата завершения или null
        type: string
      created_at:
        description: Дата создания
        example: "2026-07-08T12:00:00Z"
        type: string
      description:
        description: Описание или null
        example: 2 литра
        type: string
      id:
        description: Идентификатор задачи
        example: 1
        type: integer
      title:
        description: Заголовок
        example: Купить молоко
        type: string
      version:
        description: Версия записи
        example: 1
        type: integer
    type: object
  internal_features_tasks_transport_http.GetTaskResponse:
    properties:
      author_user_id:
        description: ID автора (из JWT)
        example: 1
        type: integer
      completed:
        description: Выполнена ли задача
        example: false
        type: boolean
      completed_at:
        description: Дата завершения или null
        type: string
      created_at:
        description: Дата создания
        example: "2026-07-08T12:00:00Z"
        type: string
      description:
        description: Описание или null
        example: 2 литра
        type: string
      id:
        description: Идентификатор задачи
        example: 1
        type: integer
      title:
        description: Заголовок
        example: Купить молоко
        type: string
      version:
        description: Версия записи
        example: 1
        type: integer
    type: object
  internal_features_tasks_transport_http.PatchTaskRequest:
    properties:
      completed:
        allOf:
        - $ref: '#/definitions/github_com_Slazzzer_golang-todoapp_internal_core_transport_http_types.Nullable-bool'
        description: Статус выполнения (опционально)
      description:
        allOf:
        - $ref: '#/definitions/github_com_Slazzzer_golang-todoapp_internal_core_transport_http_types.Nullable-string'
        description: Новое описание (опционально)
      title:
        allOf:
        - $ref: '#/definitions/github_com_Slazzzer_golang-todoapp_internal_core_transport_http_types.Nullable-string'
        description: Новый заголовок (опционально)
    type: object
  internal_features_tasks_transport_http.PatchTaskUserResponse:
    properties:
      author_user_id:
        description: ID автора (из JWT)
        example: 1
        type: integer
      completed:
        description: Выполнена ли задача
        example: false
        type: boolean
      completed_at:
        description: Дата завершения или null
        type: string
      created_at:
        description: Дата создания
        example: "2026-07-08T12:00:00Z"
        type: string
      description:
        description: Описание или null
        example: 2 литра
        type: string
      id:
        description: Идентификатор задачи
        example: 1
        type: integer
      title:
        description: Заголовок
        example: Купить молоко
        type: string
      version:
        description: Версия записи
        example: 1
        type: integer
    type: object
  internal_features_tasks_transport_http.TaskDTOResponse:
    properties:
      author_user_id:
        description: ID автора (из JWT)
        example: 1
        type: integer
      completed:
        description: Выполнена ли задача
        example: false
        type: boolean
      completed_at:
        description: Дата завершения или null
        type: string
      created_at:
        description: Дата создания
        example: "2026-07-08T12:00:00Z"
        type: string
      description:
        description: Описание или null
        example: 2 литра
        type: string
      id:
        description: Идентификатор задачи
        example: 1
        type: integer
      title:
        description: Заголовок
        example: Купить молоко
        type: string
      version:
        description: Версия записи
        example: 1
        type: integer
    type: object
  internal_features_users_transport_http.CreateUserRequest:
    properties:
      full_name:
        description: Полное имя (3–100 символов)
        example: Иван Иванов
        maxLength: 100
        minLength: 3
        type: string
      phone_number:
        description: Телефон в формате +7... (опционально)
        example: "+79991234567"
        maxLength: 15
        minLength: 10
        type: string
    required:
    - full_name
    type: object
  internal_features_users_transport_http.CreateUserResponse:
    properties:
      full_name:
        description: Полное имя
        example: Иван Иванов
        type: string
      id:
        description: Идентификатор пользователя
        example: 1
        type: integer
      phone_number:
        description: Телефон или null
        example: "+79991234567"
        type: string
      version:
        description: Версия записи (для оптимистичной блокировки)
        example: 1
        type: integer
    type: object
  internal_features_users_transport_http.GetUserResponse:
    properties:
      full_name:
        description: Полное имя
        example: Иван Иванов
        type: string
      id:
        description: Идентификатор пользователя
        example: 1
        type: integer
      phone_number:
        description: Телефон или null
        example: "+79991234567"
        type: string
      version:
        description: Версия записи (для оптимистичной блокировки)
        example: 1
        type: integer
    type: object
  internal_features_users_transport_http.PatchUserRequest:
    properties:
      full_name:
        allOf:
        - $ref: '#/definitions/github_com_Slazzzer_golang-todoapp_internal_core_transport_http_types.Nullable-string'
        description: Новое полное имя (опционально)
      phone_number:
        allOf:
        - $ref: '#/definitions/github_com_Slazzzer_golang-todoapp_internal_core_transport_http_types.Nullable-string'
        description: Новый телефон (опционально, null — удалить)
    type: object
  internal_features_users_transport_http.PatchUserResponse:
    properties:
      full_name:
        description: Полное имя
        example: Иван Иванов
        type: string
      id:
        description: Идентификатор пользователя
        example: 1
        type: integer
      phone_number:
        description: Телефон или null
        example: "+79991234567"
        type: string
      version:
        description: Версия записи (для оптимистичной блокировки)
        example: 1
        type: integer
    type: object
  internal_features_users_transport_http.UserDTOResponse:
    properties:
      full_name:
        description: Полное имя
        example: Иван Иванов
        type: string
      id:
        description: Идентификатор пользователя
        example: 1
        type: integer
      phone_number:
        description: Телефон или null
        example: "+79991234567"
        type: string
      version:
        description: Версия записи (для оптимистичной блокировки)
        example: 1
        type: integer
    type: object
host: localhost:5050
info:
  contact: {}
  description: |-
    REST API для todo-приложения (задачи, пользователи, статистика).
    Открытый API без авторизации: все эндпоинты доступны под /api/v1.
  title: Todoapp API
  version: "1.0"
paths:
  /health:
    get:
      description: Возвращает ok, если HTTP-сервер запущен. Не требует авторизации.
      produces:
      - application/json
      responses:
        "200":
          description: Сервер работает
          schema:
            $ref: '#/definitions/internal_core_transport_http_probes.ProbeStatusResponse'
      summary: Проверка живости
      tags:
      - probes
  /ready:
    get:
      description: Возвращает ready, если PostgreSQL доступен. Не требует авторизации.
      produces:
      - application/json
      responses:
        "200":
          description: Приложение готово принимать запросы
          schema:
            $ref: '#/definitions/internal_core_transport_http_probes.ProbeStatusResponse'
        "503":
          description: База данных недоступна
          schema:
            $ref: '#/definitions/internal_core_transport_http_probes.ProbeStatusResponse'
      summary: Проверка готовности
      tags:
      - probes
  /statistics:
    get:
      description: 'Агрегированная статистика по задачам. Параметр user_id опционален:
        без него — по всем пользователям.'
      parameters:
      - description: Фильтр по пользователю
        in: query
        name: user_id
        type: integer
      - description: Начало периода (YYYY-MM-DD)
        in: query
        name: from
        type: string
      - description: Конец периода (YYYY-MM-DD, не включая этот день)
        in: query
        name: to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Статистика
          schema:
            $ref: '#/definitions/internal_features_statistics_transport_http.GetStatisticsResponse'
        "400":
          description: Невалидный период или параметры
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Статистика
      tags:
      - statistics
  /tasks:
    get:
      description: Возвращает задачи. Можно отфильтровать по user_id. Поддерживает
        пагинацию limit/offset (по умолчанию limit=50, max=100).
      parameters:
      - description: Фильтр по автору
        in: query
        name: user_id
        type: integer
      - description: Размер страницы
        in: query
        name: limit
        type: integer
      - description: Смещение
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Список задач
          schema:
            items:
              $ref: '#/definitions/internal_features_tasks_transport_http.TaskDTOResponse'
            type: array
        "400":
          description: Невалидные параметры
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Список задач
      tags:
      - tasks
    post:
      consumes:
      - application/json
      description: Создаёт новую задачу с указанным автором (author_user_id).
      parameters:
      - description: Данные новой задачи
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_features_tasks_transport_http.CreateTaskRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Задача создана
          schema:
            $ref: '#/definitions/internal_features_tasks_transport_http.CreateTaskResponse'
        "400":
          description: Невалидные данные
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Создать задачу
      tags:
      - tasks
  /tasks/{id}:
    delete:
      parameters:
      - description: ID задачи
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: Задача удалена
        "404":
          description: Задача не найдена
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Удалить задачу
      tags:
      - tasks
    get:
      description: Возвращает задачу по её идентификатору.
      parameters:
      - description: ID задачи
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Данные задачи
          schema:
            $ref: '#/definitions/internal_features_tasks_transport_http.GetTaskResponse'
        "404":
          description: Задача не найдена
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить задачу
      tags:
      - tasks
    patch:
      consumes:
      - application/json
      description: Частичное обновление. При completed=true автоматически проставляется
        completed_at.
      parameters:
      - description: ID задачи
        in: path
        name: id
        required: true
        type: integer
      - description: Поля для обновления
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_features_tasks_transport_http.PatchTaskRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Обновлённая задача
          schema:
            $ref: '#/definitions/internal_features_tasks_transport_http.PatchTaskUserResponse'
        "400":
          description: Невалидные данные
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Задача не найдена
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Изменить задачу
      tags:
      - tasks
  /users:
    get:
      description: Возвращает всех пользователей. Поддерживает пагинацию limit/offset
        (по умолчанию limit=50, max=100).
      parameters:
      - description: Размер страницы
        in: query
        name: limit
        type: integer
      - description: Смещение
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Список пользователей
          schema:
            items:
              $ref: '#/definitions/internal_features_users_transport_http.UserDTOResponse'
            type: array
        "400":
          description: Невалидные параметры
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Список пользователей
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Регистрирует нового пользователя. Телефон опционален.
      parameters:
      - description: Данные нового пользователя
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_features_users_transport_http.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Пользователь создан
          schema:
            $ref: '#/definitions/internal_features_users_transport_http.CreateUserResponse'
        "400":
          description: Невалидные данные
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Создать пользователя
      tags:
      - users
  /users/{id}:
    delete:
      description: Удаляет пользователя по id. Если есть связанные задачи, может вернуться
        409.
      parameters:
      - description: ID пользователя
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: Пользователь удалён
        "404":
          description: Пользователь не найден
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Конфликт (например, есть связанные данные)
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Удалить пользователя
      tags:
      - users
    get:
      description: Возвращает данные пользователя по его идентификатору.
      parameters:
      - description: ID пользователя
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Данные пользователя
          schema:
            $ref: '#/definitions/internal_features_users_transport_http.GetUserResponse'
        "404":
          description: Пользователь не найден
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить пользователя
      tags:
      - users
    patch:
      consumes:
      - application/json
      description: Частичное обновление профиля пользователя.
      parameters:
      - description: ID пользователя
        in: path
        name: id
        required: true
        type: integer
      - description: Поля для обновления
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_features_users_transport_http.PatchUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Обновлённый профиль
          schema:
            $ref: '#/definitions/internal_features_users_transport_http.PatchUserResponse'
        "400":
          description: Невалидные данные
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Пользователь не найден
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Изменить пользователя
      tags:
      - users
swagger: "2.0"